DRYs up some graphql queries and updates relevant components
