+++ resources/assets/components/blocks/ScholarshipReferralVoterRegistrationBlock/ScholarshipReferralVoterRegistrationBlock.js
@@ -1,28 +1,12 @@
-import { get } from 'lodash';
-import gql from 'graphql-tag';
 import PropTypes from 'prop-types';
-import { useQuery } from 'react-apollo';
 import React, { Fragment } from 'react';
 
 import Card from '../../utilities/Card/Card';
-import { env, report } from '../../../helpers';
-import ErrorBlock from '../ErrorBlock/ErrorBlock';
 import DoSomethingLogo from '../../utilities/DoSomethingLogo/DoSomethingLogo';
-import PlaceholderText from '../../utilities/PlaceholderText/PlaceholderText';
+import ScholarshipInfoBlockTitle from '../../utilities/ScholarshipInfoBlockTitle/ScholarshipInfoBlockTitle';
 import StartVoterRegistrationForm from '../../utilities/StartVoterRegistrationForm/StartVoterRegistrationForm';
 
 import '../ScholarshipInfoBlock/scholarshipInfoBlock.scss';
-
-/**
- * The GraphQL query to load data for this component.
- */
-const SCHOLARSHIP_AFFILIATE_QUERY = gql`
-  query ScholarshipAffiliateQuery($utmLabel: String!, $preview: Boolean!) {
-    affiliate(utmLabel: $utmLabel, preview: $preview) {
-      title
-    }
-  }
-`;
 
 const ScholarshipReferralVoterRegistrationBlock = ({
   affiliateSponsors,
@@ -30,24 +14,6 @@ const ScholarshipReferralVoterRegistrationBlock = ({
   campaignId,
   utmLabel,
 }) => {
-  const { loading, error, data } = useQuery(SCHOLARSHIP_AFFILIATE_QUERY, {
-    skip: utmLabel === null,
-    variables: {
-      utmLabel,
-      preview: env('CONTENTFUL_USE_PREVIEW_API'),
-      campaignId,
-    },
-  });
-
-  const isLoaded = !loading;
-  const affiliateTitle = get(data, 'affiliate.title');
-
-  if (error) {
-    console.error(`[ErrorBlock] ${error}`);
-    report(error);
-    return <ErrorBlock error={error} />;
-  }
-
   return (
     <Card attributes={attributes} className="flex flex-col-reverse md:flex-row">
       <div className="md:w-1/2 p-6 mx-2 md:mx-4">
@@ -71,17 +37,10 @@ const ScholarshipReferralVoterRegistrationBlock = ({
           ) : null}
         </div>
         <div className="pt-6 pb-8 clear-both">
-          {isLoaded ? (
-            <strong className="text-lg">
-              Welcome
-              {affiliateTitle
-                ? ` from ${affiliateTitle.toUpperCase()}`
-                : ' to DoSomething.org!'}
-              !
-            </strong>
-          ) : (
-            <PlaceholderText size="large" />
-          )}{' '}
+          <ScholarshipInfoBlockTitle
+            campaignId={campaignId}
+            utmLabel={utmLabel}
+          />{' '}
           <Fragment>
             <p>
               Before you take the quiz, make sure you’re registered to vote!
